refactor(api v3): Move files and functions to allow v3 api
