+++ _requests/get/byid.ts
@@ -1,8 +1,8 @@
 // FIXME: Request could be done from the frontend
 import { VercelRequest, VercelResponse } from "@vercel/node";
-import { supabase } from "../../../_utils/supabase";
-import { setupResponseData } from "../../../_utils/setup-response";
-import { checkDataError } from "../../../_utils/data-error-response";
+import { supabase } from "../../_utils/supabase";
+import { setupResponseData } from "../../_utils/setup-response";
+import { checkDataError } from "../../_utils/data-error-response";
 
 export default async function handler(
 	request: VercelRequest,
