+++ _requests/get/lastwatered.ts
@@ -3,14 +3,14 @@ import { VercelRequest, VercelResponse } from "@vercel/node";
 import {
 	checkLimitAndOffset,
 	getLimitAndOffeset,
-} from "../../../_utils/limit-and-offset";
-import { getRange } from "../../../_utils/parse-content-range";
-import { setupResponseData } from "../../../_utils/setup-response";
-import { supabase } from "../../../_utils/supabase";
-import { getEnvs } from "../../../_utils/envs";
-import { checkRangeError } from "../../../_utils/range-error-response";
-import { checkDataError } from "../../../_utils/data-error-response";
-import { createLinks } from "../../../_utils/create-links";
+} from "../../_utils/limit-and-offset";
+import { getRange } from "../../_utils/parse-content-range";
+import { setupResponseData } from "../../_utils/setup-response";
+import { supabase } from "../../_utils/supabase";
+import { getEnvs } from "../../_utils/envs";
+import { checkRangeError } from "../../_utils/range-error-response";
+import { checkDataError } from "../../_utils/data-error-response";
+import { createLinks } from "../../_utils/create-links";
 const { SUPABASE_URL } = getEnvs();
 
 export default async function handler(
