+++ api/post/[type].ts
@@ -1,23 +1,16 @@
 import { VercelRequest, VercelResponse } from "@vercel/node";
 import setHeaders from "../../_utils/set-headers";
-import { supabase } from "../../_utils/supabase";
 import { postSchemas, validate } from "../../_utils/validation";
-import { Database } from "../../_types/database";
-import { verifyRequest } from "../../_utils/verify";
 import { queryTypes as queryTypesList } from "../../_utils/routes-listing";
+import { verifyAuth0Request } from "../../_utils/verify-auth0";
+import adoptHandler from "../../_requests/post/adopt";
+import waterHandler from "../../_requests/post/water";
 
 const queryTypes = Object.keys(queryTypesList["POST"]);
 
 // api/[name].ts -> /api/lee
 // req.query.name -> "lee"
 
-// const schemas: Record<string, AjvSchema> = {
-// 	adopt: adoptSchema,
-// 	water: waterSchema,
-// };
-
-// type TreesAdopted = Database["public"]["Tables"]["trees_adopted"]["Insert"];
-type TreesWatered = Database["public"]["Tables"]["trees_watered"]["Insert"];
 export default async function postHandler(
 	request: VercelRequest,
 	response: VercelResponse
@@ -26,6 +19,7 @@ export default async function postHandler(
 	if (request.method === "OPTIONS") {
 		return response.status(200).end();
 	}
+
 	// const { data: userData, error } = await verifySupabaseToken(request);
 	// if (error) {
 	// 	console.error("error from supabase auth", error);
@@ -38,6 +32,8 @@ export default async function postHandler(
 	/**
 	 * We will remove auth0 but for now we can auth with both
 	 */
+	const auth0RequestValid = await verifyAuth0Request(request);
+	if (!auth0RequestValid) {
 		return response.status(401).json({ error: "unauthorized" });
 	}
 	const { type } = request.query;
@@ -66,43 +62,10 @@ export default async function postHandler(
 		// https://github.com/technologiestiftung/giessdenkiez-de-postgres-api/issues/159
 
 		case "adopt": {
-			const { tree_id, uuid } = request.body;
-			const { data, error } = await supabase
-				.from("trees_adopted")
-				.upsert(
-					{
-						tree_id,
-						uuid,
-					},
-
-					{ onConflict: "uuid,tree_id" }
-				)
-				.select();
-			if (error) {
-				// console.error(error);
-				return response.status(500).json({ error });
-			}
-			return response.status(201).json({ message: "adopted", data });
+			return await adoptHandler(request, response);
 		}
 		case "water": {
-			const body = request.body as TreesWatered;
-			const { tree_id, username, timestamp, uuid, amount } = body;
-			const { data, error } = await supabase
-				.from("trees_watered")
-				.insert({
-					// TODO: [GDK-220] Remove time from db schema trees_watered  it is a legacy value not used anymore
-					// https://github.com/technologiestiftung/giessdenkiez-de-postgres-api/issues/160
-					tree_id,
-					username,
-					timestamp,
-					uuid,
-					amount,
-				})
-				.select();
-			if (error) {
-				return response.status(500).json({ error });
-			}
-			return response.status(201).json({ message: "watered", data });
+			return await waterHandler(request, response);
 		}
 	}
 }
